:boom: Migrating to Bootsrap
Moving away from Bulma to Bootstrap.

## Why?
As the first focus is more about the UX and less about the UI itself, using a library like `reactstrap` help speeding the develop process with a more battle tested library to handle the UI.
`reactstrap` having way more stars than `react-bulma`, and even if I was not using `react-bulma`, not developing my own components helps.

## Changes
I removed packages for Bulma and for the `sass`, then added the `reactstrap` packages and the `@zeit/next-css` one to add to the next config.