:sparkles: Create Nav component
Delete the `nav` component that was on the nextjs boilerplate, and create a proper Tailwind one.

The `Nav` is populated with some links (useless for now, but will be useful in the future once we add more pages).
It use a state only on mobile to display or not the menu.