:sparkles: Create a Button and Spinner component
The `Button` component work like a classic `button` tag.
By having a set button component, any futur part of the project that needs a button will be consistent in design as well as behavior.

The `Spinner` is a nice css spinner to indicate loading. So far it’s only use on the `Button`.