:construction: Write draft on Plex auth
Based on the read there[0], I started implementing a draft of a Plex auth.

On the login page, Kino create a request to get a code and redirect the user to the Plex auth page. In the Plex request, a callback is added to a new page of Kino (the `redirect` page), which will get a code.

Once on the `redirect` page, the code is used to query a `authToken` from Plex on behalf of the user. And with that token we can query the user itself, which could be used later to authenticate it on Kino.

So far we only show the user data to him/herself on the `redirect` page.

[0] https://github.com/tidusjar/Ombi/issues/2894#issuecomment-477404691