:recycle: Change dotenv by a config file
First, for now the values used in `.env` were only used on the frontend side so they were not that secrets.

Then `now` don’t setup the secrets on the front, only on the server, so those vars were useless in production.

So to fix that, I just put the env in a `.ts` file.
Gud enough