:sparkles: Create a component Card
In addition of creating a component `Card`, we also might need to extend the style of the `Card`.
To do so more easily, I created a type `KFC`, which extends the FunctionComponent from React, and add an optional `className`.

`KFC` stand for `KinoFunctionComponent`.

The type for `KFC` is a new @types folder that will hold more type or this kind.