:lipstick: Fix rounded corner for MovieCard
The rounded corner was added to the `img` and not the parent, and as the parent had a background color, a slight part of the background was shown on the edge of the corners.

So the solution was to put the rounded corner on the parent `div` and set the `overflow` to hidden.