:sparkles: Create a useDebounce hook
Add libraries `use-constant` and `awesome-debounce-promise`, those are use to create the `useDebounce` hook. Those packages are taken from this StackOverflow answer[0]. The `react-async-hook` is not necessary as `react-query` provide the same end result.

The main reason of creating that hook is to provide a clear and easy way to debounce any requests that needs to be (also doing that in a hook way is a nice bonus). It’s now used in the `index.tsx` for the movie/tv search so you can see how it can be used.

[0] https://stackoverflow.com/questions/23123138/perform-debounce-in-react-js/28046731#28046731