:recycle: Make the Spinner a KFC
Make the `Spinner` a `KFC` type, so it can get its `className` overrided.
As the first change in that, remove the `margin: auto` from the css as it’s should be handled by tailwind instead of the spinner itself.

And to fix that UI change on the `Button`, the `span` around the spinner is now useless as the className can be added directly in the `Spinner`. I also added the `items-center` to make sure the `Spinner` is align with the button’s children.