:lipstick: Add a loading indicator for search
During the movie/tv search, I added a Spinner to show the loading of data.

To do so, I created a `searched` state that save if the user has typed at least once a key. If so, the Results label will be shown for the rest of the page use.
And when it’s loading a spinner aligned with the baseline is displayed.