:bug: Fix axios clients because react-query types
As I added the types for `react-query`, the response for the use of `react-query` is now expecting a full `AxiosResponse` object. Which was not the case as (thanks to interceptors) only the `data` from the response was return.

So this commit just revert that. So now the types are happy